pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"fastapi>=0.115.0",
	"uvicorn[standard]>=0.30.0",
	"sqlalchemy>=2.0.0",
	"asyncpg>=0.29.0",
	"passlib[bcrypt]>=1.7.4",
	"bcrypt<5",
	"python-jose[cryptography]>=3.3.0",
	"python-dotenv>=1.0.1",
	"email-validator>=2.2.0",
]

[project.scripts]
fastapi-ready = "fastapi_ready:main"

[tool.setuptools]
py-modules = ["main", "auth", "deps", "models", "schemas", "fastapi_ready"]

[tool.setuptools.packages.find]
include = ["config*", "db*", "router*"]
exclude = ["auth*", "core*"]
